refactor and add validation for max level 10
